解决只能编译一个模块 .cs 文件的问题

+++ ObjectServer/ObjectServer.Test/modules/test/test-object.cs
@@ -86,7 +86,7 @@ namespace ObjectServer.Test
                 }
                 else
                 {
-                    rows[id] = (long)field1 + (long)field2;
+                    rows[id] = (int)field1 + (int)field2;
                 }
             }
             return rows;
